refactor: migrate disruption and buildkit commands from bash scripts to CLI
- Update kube_disruption_window_controller to use pf k8s disruptions enable/disable
- Update kube_buildkit to use pf buildkit clear-cache and scale down commands
- Fix missing warning when disruption window seconds annotation is absent
- Add container runtime integration documentation to CLAUDE.md
- Document CLI migration patterns for converting legacy scripts

All commands maintain functional parity with original bash scripts while
providing better error handling and user feedback through the CLI framework.
