+++ packages/infrastructure/kube_buildkit/main.tf
@@ -358,9 +358,8 @@ module "scale_to_zero" {
     image_registry   = "public.ecr.aws"
     image_repository = module.constants.panfactum_image_repository
     image_tag        = module.constants.panfactum_image_tag
-    // todo: replace with pf buildkit
     command = [
-      "/bin/pf-buildkit-scale-down",
+      "pf", "buildkit", "scale", "down",
       "--timeout",
       tostring(var.scale_down_delay_seconds)
     ]
@@ -438,9 +437,8 @@ module "cache_clear" {
     image_repository = module.constants.panfactum_image_repository
     image_tag        = module.constants.panfactum_image_tag
 
-    // todo: replace with pf buildkit cache clear
     command = [
-      "/bin/pf-buildkit-clear-cache",
+      "pf", "buildkit", "clear-cache"
     ]
     minimum_memory = 50
   }]
