+++ packages/infrastructure/kube_disruption_window_controller/main.tf
@@ -90,9 +90,8 @@ module "disruption_window_enabler" {
     image_registry   = "public.ecr.aws"
     image_repository = module.constants.panfactum_image_repository
     image_tag        = module.constants.panfactum_image_tag
-    // todo: implement pf k8s disruptions enable
     command = [
-      "/bin/pf-voluntary-disruptions-enable",
+      "pf", "k8s", "disruptions", "enable",
       "--window-id=${random_id.window_id.hex}",
       "--namespace=${var.namespace}"
     ]
@@ -120,9 +119,8 @@ module "disruption_window_disabler" {
     image_registry   = "public.ecr.aws"
     image_repository = module.constants.panfactum_image_repository
     image_tag        = module.constants.panfactum_image_tag
-    // todo: implement pf k8s disruptions disable
     command = [
-      "/bin/pf-voluntary-disruptions-disable",
+      "pf", "k8s", "disruptions", "disable",
       "--window-id=${random_id.window_id.hex}",
       "--namespace=${var.namespace}"
     ]
